cleaning up

+++ echo.c
@@ -12,7 +12,7 @@ void echo(int connfd)
     
     Rio_readinitb(&rio, connfd);
     while((n = Rio_readlineb(&rio, buf, MAXLINE)) != 0) { //line:netp:echo:eof
-        printf("server received %d bytes\n", n);
+        printf("server received %d bytes\n", (int) n);
         Rio_writen(connfd, buf, n);
     }
 }
